some cosmetic and debugging
configfiles will be read in from logscan and no longer from logscanlib.
(It's the job of the prog, not of the lib.)
timecodes can be passed via logscanlib.add_timecodes

logscan: files will be closed in finally-clause

section includes start- AND end-time.
Before end-time was not included. (With the stupid side effect that the
last line of the log often was missing; logscan took the log.end-time as
end-time for section, so if this wouldn't be included...)
But this could be solved otherwise. It just seems more intuitive to
include the end-time. (Does it? I'm not sure...)